Add files via upload
In this project, we will go through 8 main questions:

In how many cities Summer Olympics is held so far?
Which sport is having the most number of Gold medals so far? (Top 5 in graph format)
Which sport is having the most number of medals so far? (Top 5 in graph format)
Which player has won the most number of medals? (Top 5 in graph format)
Which player has won the most number of Gold Medals medals? (Top 5 in graph format)
In which year India won the first Gold Medal in Summer Olympics?
Which event is most popular in terms of the number of players? (Top 5 in graph format)
Which sport is having the most female Gold Medalists? (Top 5 in graph format)